# api-service/.artifacts/openapi-billing.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent SaaS API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/auth/request-access": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Submit Access Request",
        "operationId": "submit_access_request_api_v1_auth_request_access_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequestPublicRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupAccessPolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/signup-requests": {
      "get": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "List Signup Requests",
        "operationId": "list_signup_requests_api_v1_auth_signup_requests_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SignupRequestStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupRequestListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/signup-requests/{request_id}/approve": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Approve Signup Request",
        "operationId": "approve_signup_request_api_v1_auth_signup_requests__request_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequestApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupRequestDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/signup-requests/{request_id}/reject": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Reject Signup Request",
        "operationId": "reject_signup_request_api_v1_auth_signup_requests__request_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequestRejectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupRequestDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SignupAccessPolicyResponse": {
        "properties": {
          "policy": {
            "type": "string",
            "title": "Policy",
            "description": "Active signup access policy."
          },
          "invite_required": {
            "type": "boolean",
            "title": "Invite Required",
            "description": "Indicates whether an invite token is required."
          },
          "request_access_enabled": {
            "type": "boolean",
            "title": "Request Access Enabled",
            "description": "Whether deployments expose the access-request workflow."
          }
        },
        "type": "object",
        "required": [
          "policy",
          "invite_required",
          "request_access_enabled"
        ],
        "title": "SignupAccessPolicyResponse"
      },
...
      "SignupInviteIssueResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "token_hint": {
            "type": "string",
            "title": "Token Hint"
          },
          "invited_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invited Email"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "max_redemptions": {
            "type": "integer",
            "title": "Max Redemptions"
          },
          "redeemed_count": {
            "type": "integer",
            "title": "Redeemed Count"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "signup_request_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signup Request Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "invite_token": {
            "type": "string",
            "title": "Invite Token",
            "description": "Plaintext invite token (only returned once)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "token_hint",
          "status",
          "max_redemptions",
          "redeemed_count",
          "created_at",
          "invite_token"
        ],
        "title": "SignupInviteIssueResponse"
      },
...
      "SignupRequestApprovalRequest": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Optional note recorded alongside the invite and decision."
          },
          "invite_expires_in_hours": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 336.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Invite Expires In Hours",
            "description": "Optional invite expiry override in hours.",
            "default": 72
          }
        },
        "type": "object",
        "title": "SignupRequestApprovalRequest"
      },
      "SignupRequestDecisionResponse": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/SignupRequestResponse"
          },
          "invite": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignupInviteIssueResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "request"
        ],
        "title": "SignupRequestDecisionResponse"
      },
      "SignupRequestListResponse": {
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/SignupRequestResponse"
            },
            "type": "array",
            "title": "Requests"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "requests",
          "total"
        ],
        "title": "SignupRequestListResponse"
      },
      "SignupRequestPublicRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "organization": {
            "type": "string",
            "maxLength": 128,
            "minLength": 2,
            "title": "Organization"
          },
          "full_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 2,
            "title": "Full Name"
          },
          "message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Optional context describing the intended use case."
          },
          "accept_terms": {
            "type": "boolean",
            "title": "Accept Terms",
            "description": "Indicates that the requester accepted the Terms."
          },
          "honeypot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Honeypot",
            "description": "Hidden honeypot field used to detect basic bots."
          }
        },
        "type": "object",
        "required": [
          "email",
          "organization",
          "full_name",
          "accept_terms"
        ],
        "title": "SignupRequestPublicRequest"
      },
      "SignupRequestRejectionRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Reason communicated to operators when rejecting a request."
          }
        },
        "type": "object",
        "title": "SignupRequestRejectionRequest"
      },
      "SignupRequestResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "invite_token_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invite Token Hint"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "organization",
          "full_name",
          "status",
          "created_at"
        ],
        "title": "SignupRequestResponse"
      },
      "SignupRequestStatus": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "rejected"
        ],
        "title": "SignupRequestStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}